Record-only MFC of r330252, r330253
r330253 reverted r330252, which was committed with nothing but the message:
"MFC After: 1 week"

Go ahead and record the merge of both of them. r330254 was the follow-up
commit with the same contents as r330252 but with an appropriate message. It
will be MFC'd along with the introduction of the corresponding options as
appropriate.


git-svn-id: https://svn.freebsd.org/base/stable/11@331685 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f