MFC r325422: posix_fallocate.2: add an EINVAL errno case
As of r325320 in HEAD posix_fallocate returns EINVAL on ZFS to indicate
that the underlying filesystem does not support this operation, per
POSIX.1-2008. Document this case in the man page.

Note that r325320 has not yet been merged to stable/11, and may or may
not be.  However, we should document that EINVAL may be returned if the
filesystem does not support posix_fallocate (even if we don't actually
do so in stable/11), as software should be prepared to handle that case.

Discussed with:	avg
Sponsored by:	The FreeBSD Foundation


git-svn-id: https://svn.freebsd.org/base/stable/11@331687 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f