MFC r329373: Correct module symbol export handling
EXPORT_SYMS can be set to YES, NO, a list of symbols to export from a
module, or to a filename containing such a list.  For the case that it
is set to a symbol list, replace spaces in the list with newlines, so
the created file is in the format expected by kmod_syms.awk.

Sponsored by:	Turing Robotic Industries Inc.


git-svn-id: https://svn.freebsd.org/base/stable/11@331751 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f