MFC r312378 by andrew:
Use the kernel stack in the ARM FBT DTrace provider. This is used to find
the fifth argument to functions being traced, however there was an error
where the userspace stack was being used. This may be invalid leading to
a kernel panic if this address is unmapped.

Submitted by:	Graeme Jenkinson <graeme.jenkinson@cl.cam.ac.uk>
Differential Revision:	https://reviews.freebsd.org/D9229

MFC commit fields
PR:		211389


git-svn-id: https://svn.freebsd.org/base/stable/11@331952 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f