netmap: align if_ptnet guest driver to the upstream code (commit 0e15788)
The change upgrades the driver to use the split Communication Status
Block (CSB) format. In this way the variables written by the guest
and read by the host are allocated in a different cacheline than
the variables written by the host and read by the guest; this is
needed to avoid cache thrashing.

Approved by:	hrs (mentor)


git-svn-id: https://svn.freebsd.org/base/head@332047 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f