Record-only MFC of r325365, r325564
r325365 fixed a bug in patch(1) that would cause a match if we ran out of
context, say, at the beginning/end of a file. r325564 reverted it at the
request of antoine until an exp-run could be completed to fix all of the
ports with broken patches.

r325365 was later committed as r326084, which has already been MFC'd.


git-svn-id: https://svn.freebsd.org/base/stable/11@332051 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f