MFC r331718: stand: Add workaround for HP BIOS issues
hrs@ and kuriyama@ have found that on some HP BIOS, a system will fail to
boot immediately after installation with the claim that it can't work out
which disk they are booting from.

They tracked it down to a buffer overrun, and found that it could be
alleviated by doing a dummy read before-hand.


git-svn-id: https://svn.freebsd.org/base/stable/11@332056 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f