stable/11: Bump FreeBSD_Version for recent rc ${name}_limits bits
Some ports had their own interpretation of _limits that conflicted with the
new definition as of r328331. r331880 was an MFC of this commit to
stable/11. Bump FreeBSD_Version in stable/11 following r331880 so that ports
can do the right thing.

This is a direct commit to stable/11

Reported by:	0mp


git-svn-id: https://svn.freebsd.org/base/stable/11@332089 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f