lualoader: Fix menu skipping with loader.conf(5) vars
Earlier efforts to stop loading the menu broke the ability to skip the menu
with, e.g., beastie_disable in loader.conf(5) as it was decided before
configuration was read.

Defer bringing in the menu module until we've loaded configuration so that
we can make a more informed decision on whether the menu should be skipped
or not.


git-svn-id: https://svn.freebsd.org/base/head@332106 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f