MFC r329859,r329860: Float protection in stand
r329859: Do not include float interfaces when using libsa.

We don't support float in the boot loaders, so don't include
interfaces for float or double in systems headers. In addition, take
the unusual step of spiking double and float to prevent any more
accidental seepage.

r329860: Floaty McFloatface is funnier...


git-svn-id: https://svn.freebsd.org/base/stable/11@332135 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f