MFC r324638:
Fix the handling of parital and too short chunks.

Ensure that the current behaviour is consistent: stop processing
of the chunk, but finish the processing of the previous chunks.

This behaviour might be changed in a later commit to ABORT the
assoication due to a protocol violation, but changing this
is a separate issue.

MFC r324725

Fix a bug introduced in r324638.
Thanks to Felix Weinrank for making me aware of this.

MFC r324726:

Revert change which got in accidently.


git-svn-id: https://svn.freebsd.org/base/stable/11@332213 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f