MFC r327844:
Ensure that the vnet is set when calling pru_sockaddr() and
pru_peeraddr().

This is already true when called via kern_getsockname() and
kern_getpeername(). This patch sets it also, when they arecalled
via soo_fill_kinfo(). This is necessary, since the corresponding
functions for SCTP require the vnet to be set. Without this,
if a process having an wildcard bound SCTP socket is
terminated and a core is written, the kernel panics.


git-svn-id: https://svn.freebsd.org/base/stable/11@332230 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f