MFC r328066:
Fix a bug related to fast retransmissions.

When processing a SACK advancing the cumtsn-ack in fast recovery,
increment the miss-indications for all TSN's reported as missing.

Thanks to Fabian Ising for finding the bug and to Timo Voelker
for provinding a fix.

This fix moves also CMT related initialisation of some variables
to a more appropriate place.


git-svn-id: https://svn.freebsd.org/base/stable/11@332234 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f