MFC r331654, r331869
r331654:
Don't access userspace directly from the kernel in nxge(4).

Update to what the previous code seemed to be doing via the correct
interfaces.  Further issues exist in xge_ioctl_registers(), but this is
debugging code in a driver that has few users and they don't appear to
be crashes or leaks.

Reviewed by:	jhb (prior version)
Sponsored by:	DARPA, AFRL
Differential Revision:	https://reviews.freebsd.org/D14848

r331869:
Fix the build on arches with default unsigned char.  Capture the fubyte()
return value in an int as well as the char, and test the full int value
for fubyte() failure.


git-svn-id: https://svn.freebsd.org/base/stable/10@332280 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f