MFC ath(4) potential memory disclosure fixes
[1] r327499: ath: fix memory disclosure from ath_btcoex_ioctl

The ath_btcoex_ioctl handler allocated a buffer without M_ZERO and
returned it to userland without writing to it.

The device has permissions only for root so this is not urgent, and the
fix can be MFCd and considered for a future EN.

[2] r327500: ath: fix possible memory disclosures in ioctl handlers

Apply the fix from r327499 to additional ioctl handlers.

Note: related fix in r327529 does not apply directly to stable/10 and
will be addressed in a followup commit.

Submitted by:	Domagoj Stolfa <domagoj.stolfa@gmail.com> [1]
Reported by:	Ilja van Sprundel <ivansprundel@ioactive.com> [1,2]
Reviewed by:	adrian [1]
Sponsored by:	The FreeBSD Foundation


git-svn-id: https://svn.freebsd.org/base/stable/10@332320 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f