Revert r331880, MFC of r328331 and bump FreeBSD_version
There are logistics issues that weren't considered when this was originally
MFC'd. All rc scripts in ports need audited (this is in progress) for usage
of ${name}_limits that doesn't line up with the new interpretation, and
individual rc.conf(5)'s need to be scrubbed of usage that doesn't line up.

It's since been decided that it should be left for a feature in 12.

1101514 introduced interpretation of ${name}_limits for rc scripts; this
feature no longer exists as of 1101515.


git-svn-id: https://svn.freebsd.org/base/stable/11@332363 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f