Refactor currdev setting
Refactor the currdev setting to find the device we booted from. Limit
searching when we don't already have a reasonable currdev from that to
the same device only. Search a little harder for ZFS volumes as that's
needed for loader.efi to live on an ESP.

Sponsored by: Netflix
Differential Review: https://reviews.freebsd.org/D13784


git-svn-id: https://svn.freebsd.org/base/head@332416 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f