MFC r331077 (brooks): Add _IOC_NEWLEN() and _IOC_NEWTYPE() macros.
These macros take an existing ioctl(2) command and replace the length
with the specified length or length of the specified type respectively.
These can be used to define commands for 32-bit compatibility with fewer
opportunities for cut-and-paste errors then a whole new definition.

Obtained from:	CheriBSD
Sponsored by:	DARPA, AFRL


git-svn-id: https://svn.freebsd.org/base/stable/11@332502 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f