MFC r328589:
Make the handler routine for the hw.usb.template sysctl trigger the USB
host to reprobe the bus by switching the USB pull up resistors off and
back on.  In other words - when FreeBSD is configured as a USB device,
changing the sysctl will be immediately noticed by the machine it's
connected to.

Relnotes:	yes
Sponsored by:	The FreeBSD Foundation


git-svn-id: https://svn.freebsd.org/base/stable/11@332598 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f