MFC r325316:
Add benchmark for getresuid(2) - three pointers, although only "output" ones,
and quite lightwait.

The purpose of this commit, and the previous one, is to be able to measure
overhead of pointer arguments - in case you're running a strange architecture
where pointers and integers are quite different things at the hardware level.


git-svn-id: https://svn.freebsd.org/base/stable/11@332626 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f