MFC r332561,r332573: Rename volume label for ESP
MFC r332561: Rename volume label for ESP

Harry Schmalzbauer reports that some firmware, in his experience, trips
over the ESP we install due to the volume label. It has been theorized that
this is due to some confusion with the label and the path on the ESP to
boot1.efi.

Regardless, Harry found that renaming the label seems to fix it.

MFC r332573: Regenerate FAT templates after r332561

PR:		214282


git-svn-id: https://svn.freebsd.org/base/stable/11@332746 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f