MFC r332429, r332431
r332429:
cron(8): Reload database if an existing job in cron.d changed as well

Directory mtime will only change if a file is added or removed, not
modified. For /var/cron/tabs, this is fine because of how crontab(1) manages
it using temp files so all crontab(1) changes will trigger a reload of the
database.

For /etc/cron.d and /usr/local/etc/cron.d, this is not necessarily the case.
Instead of checking their mtime, we should descend into them and check mtime
on all jobs also.

r332431:
cron(8): Correct test sense

We're about to use the result of fstat(2) either way, so don't do that if it
fails...


git-svn-id: https://svn.freebsd.org/base/stable/11@332747 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f