MFC r328957: {ext2|ufs}_readdir: Avoid setting negative ncookies.
ncookies cannot be negative or the allocator will fail. This should only
happen if a caller is very broken but we can still try to survive the
event.

We should probably also verify for uio_resid > MAXPHYS but in that case
it is not clear that just clipping the ncookies value is an adequate
response.


git-svn-id: https://svn.freebsd.org/base/stable/11@332749 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f